fix(api): parsing event details 'cpu_percentage'
The `cpu_percentage` field inside the `entity_map` is a `float32`
not `int32`, this commit is fixing the following error:
```
ERROR unable to get event details: json: cannot unmarshal number 0.02 into Go struct field eventProcessEntity.data.entity_map.process.cpu_percentage of type int32
```

Closes https://github.com/lacework/go-sdk/issues/78

+++ api/events.go
@@ -171,7 +171,7 @@ type eventProcessEntity struct {
 	ProcessId        int32     `json:"process_id"`
 	ProcessStartTime time.Time `json:"process_start_time"`
 	Cmdline          string    `json:"cmdline"`
-	CpuPercentage    int32     `json:"cpu_percentage"`
+	CpuPercentage    float32   `json:"cpu_percentage"`
 }
 
 type eventFileDataHashEntity struct {
